+++ src/Toggly/FeatureManagement/SdkIdentity.php
@@ -8,7 +8,7 @@ namespace Toggly\FeatureManagement;
 final class SdkIdentity
 {
     public const SDK_ID = 'php';
-    public const SDK_VERSION = '1.0.1';
+    public const SDK_VERSION = '0.2.0';
 
     private function __construct()
     {